Refactor: Remove versioning from namespaces and projects
Significant refactor of the `SFA.DAS.Payments.Audit.ArchiveService` project:
- Updated namespaces to remove versioning (`V1`).
- Adjusted using directives to match new namespaces.
- Updated project files (`.csproj`) to remove versioning.
- Modified test and implementation files to reflect new namespaces.
- Updated solution file (`.sln`) and `launchSettings.json`.
- Recreated project files to ensure consistency.

These changes simplify the structure and improve maintainability.
