+++ src/SFA.DAS.Payments.Audit.ArchiveService.UnitTests.V1/DurableTaskClientWrapper.cs
@@ -2,7 +2,7 @@
 using Microsoft.Azure.Functions.Worker.Http;
 using Microsoft.DurableTask.Client;
 
-namespace SFA.DAS.Payments.Audit.ArchiveService.UnitTests.V1
+namespace SFA.DAS.Payments.Audit.ArchiveService.UnitTests
 {
     public class DurableTaskClientWrapper : IDurableTaskClientWrapper
     {
