+++ src/SFA.DAS.Payments.Audit.ArchiveService.UnitTests.V1/Orchestrators/PeriodEndArchiveOrchestratorTests.cs
@@ -1,14 +1,14 @@
 ﻿using Microsoft.DurableTask;
 using Microsoft.Extensions.Logging;
 using Moq;
-using SFA.DAS.Payments.Audit.ArchiveService.V1.Activities;
-using SFA.DAS.Payments.Audit.ArchiveService.V1.Configuration;
-using SFA.DAS.Payments.Audit.ArchiveService.V1.Helper;
-using SFA.DAS.Payments.Audit.ArchiveService.V1.Models;
-using SFA.DAS.Payments.Audit.ArchiveService.V1.Orchestrators;
+using SFA.DAS.Payments.Audit.ArchiveService.Activities;
+using SFA.DAS.Payments.Audit.ArchiveService.Configuration;
+using SFA.DAS.Payments.Audit.ArchiveService.Helper;
+using SFA.DAS.Payments.Audit.ArchiveService.Models;
+using SFA.DAS.Payments.Audit.ArchiveService.Orchestrators;
 using SFA.DAS.Payments.Monitoring.Jobs.Messages.Commands;
 
-namespace SFA.DAS.Payments.Audit.ArchiveService.UnitTests.V1.Orchestrators
+namespace SFA.DAS.Payments.Audit.ArchiveService.UnitTests.Orchestrators
 {
     [TestFixture]
     public class PeriodEndArchiveOrchestratorTests
