+++ src/SFA.DAS.Payments.Audit.ArchiveService.UnitTests.V1/SFA.DAS.Payments.Audit.ArchiveService.UnitTests.csproj
@@ -20,7 +20,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <ProjectReference Include="..\SFA.DAS.Payments.Audit.ArchiveService.V1\SFA.DAS.Payments.Audit.ArchiveService.V1.csproj" />
+	  <ProjectReference Include="..\SFA.DAS.Payments.Audit.ArchiveService.V1\SFA.DAS.Payments.Audit.ArchiveService.csproj" />
 	</ItemGroup>
 
 	<ItemGroup>
