+++ src/SFA.DAS.Payments.Audit.ArchiveService.V1/Helper/EntityHelper.cs
@@ -1,11 +1,11 @@
 ﻿using Microsoft.DurableTask.Client;
 using Microsoft.DurableTask.Entities;
 using Microsoft.Extensions.Logging;
-using SFA.DAS.Payments.Audit.ArchiveService.V1.EntityTrigger;
+using SFA.DAS.Payments.Audit.ArchiveService.EntityTrigger;
 using SFA.DAS.Payments.Model.Core.Audit;
-using static SFA.DAS.Payments.Audit.ArchiveService.V1.Helper.StatusHelper;
+using static SFA.DAS.Payments.Audit.ArchiveService.Helper.StatusHelper;
 
-namespace SFA.DAS.Payments.Audit.ArchiveService.V1.Helper
+namespace SFA.DAS.Payments.Audit.ArchiveService.Helper
 {
     public class EntityHelper : IEntityHelper
     {
