+++ src/SFA.DAS.Payments.Audit.ArchiveService.V1/Helper/IDataFactoryHelper.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.Azure.Management.DataFactory;
 
-namespace SFA.DAS.Payments.Audit.ArchiveService.V1.Helper
+namespace SFA.DAS.Payments.Audit.ArchiveService.Helper
 {
     public interface IDataFactoryHelper
     {
